requirements.txt
asgiref==3.3.4
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
httpie==2.4.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
Markdown==3.3.4
mccabe==0.6.1
protobuf==3.15.8
pycodestyle==2.6.0
Pygments==2.9.0
pylint==2.6.0
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
wrapt==1.12.1
